requirements.txt
langchain 
boto3   
pinecone-client  
tqdm   
unstructured  
sentence-transformers   
python-magic-bin  
